Simulator/requirements.txt
azure-iot-device==2.12.0
certifi==2022.9.24
charset-normalizer==2.1.1
deprecation==2.1.0
idna==3.4
janus==1.0.0
packaging==21.3
paho-mqtt==1.6.1
pip==22.3.1
pymssql==2.2.7
pyparsing==3.0.9
PySocks==1.7.1
requests==2.28.1
requests-unixsocket==0.3.0
setuptools==41.2.0
typing_extensions==4.4.0
urllib3==1.26.13
